+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,52 +1,21 @@
 package wooteco.subway.dao;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.stereotype.Repository;
-
 import wooteco.subway.domain.Station;
 
-@Repository
-public class StationDao {
+public interface StationDao {
 
-    private static final List<Station> stations = new ArrayList<>();
-    private static Long seq = 0L;
+    Station save(Station station);
 
-    public Station save(Station station) {
-        Station persistStation = createNewObject(station);
-        stations.add(persistStation);
-        return persistStation;
-    }
+    Optional<Station> findByName(String name);
 
-    private Station createNewObject(Station station) {
-        return new Station(++seq, station.getName());
-    }
+    List<Station> findAll();
 
-    public Optional<Station> findByName(String name) {
-        return stations.stream()
-            .filter(station -> name.equals(station.getName()))
-            .findFirst();
-    }
+    void deleteById(Long id);
 
-    public List<Station> findAll() {
-        return List.copyOf(stations);
-    }
+    void deleteAll();
 
-    public void deleteById(Long id) {
-        if (!stations.removeIf(station -> station.getId().equals(id))) {
-            throw new IllegalArgumentException("존재하지 않는 역 입니다.");
-        }
-    }
-
-    public void deleteAll() {
-        stations.clear();
-    }
-
-    public Optional<Station> findById(Long id) {
-        return stations.stream()
-            .filter(station -> station.getId().equals(id))
-            .findFirst();
-    }
+    Optional<Station> findById(Long id);
 }
